feat: add onboarding card data structure and enhance localization for onboarding flow
