+++ src/app/(flow)/onboarding.tsx
@@ -3,10 +3,10 @@ import WhatsNewBox from '@/components/Flow/WhatsnewBox'
 import AnimatedText from '@/components/Flow/svgs/AnimatedText'
 import LogoSVG from '@/components/Flow/svgs/logo'
 import LogoTextSVG from '@/components/Flow/svgs/logoText'
-import PlatformIcon, { type LucideIcon } from '@/components/Universal/Icon'
+import PlatformIcon from '@/components/Universal/Icon'
 import { PRIVACY_URL } from '@/data/constants'
 import { useFlowStore } from '@/hooks/useFlowStore'
-import { type MaterialIcon } from '@/types/material-icons'
+import { type OnboardingCardData } from '@/types/data'
 import { getContrastColor } from '@/utils/ui-utils'
 import * as Haptics from 'expo-haptics'
 import { router } from 'expo-router'
@@ -41,16 +41,8 @@ export default function OnboardingScreen(): JSX.Element {
     const setAnalyticsAllowed = useFlowStore(
         (state) => state.setAnalyticsAllowed
     )
-    interface Data {
-        title: string
-        description: string
-        icon: {
-            ios: string
-            android: MaterialIcon
-            web: LucideIcon
-        }
-    }
-    const data: Data[] = [
+
+    const data: OnboardingCardData[] = [
         {
             title: t('onboarding.cards.title1'),
             description: t('onboarding.cards.description1'),
