+++ src/localization/en/flow.json
@@ -7,6 +7,8 @@
     "onboarding": {
         "links": {
             "privacy": "Privacy Policy",
+            "faq": "Learn more",
+            "imprint": "Imprint",
             "agree1": "By continuing, you agree to the ",
             "agree2": ".",
             "safety": "Data Safety"
@@ -20,7 +22,9 @@
             "title2": "Developed by students",
             "description2": "Neuland Next is an open-source project by Neuland Ingolstadt e.V. and is an alternative to the official THI app.",
             "title3": "Security comes first",
-            "description3": "Your data is strictly protected by using the official and encrypted services of THI and is never accessible to us or third parties."
+            "description3": "Your data is strictly protected by using the official and encrypted services of THI and is never accessible to us or third parties.",
+            "title4": "Even better in the app",
+            "description4": "Install Neuland Next on your smartphone to use all features. Available for iOS, iPadOS, MacOS, and Android."
         }
     },
     "login": {
